+++ src/test/java/com/redhat/labs/omp/resources/GitSyncResourceTest.java
@@ -236,80 +236,6 @@ public class GitSyncResourceTest {
 
     }
 
-    /*
-     * GIT 2 Backend Refresh
-     *  - create 1, then run sync
-     *  - git api returns 1 or 2 engagements, updates mongo 
-     *  - do get all (should match what is returned from git api)
-     */
-    @Test
-    public void testRefreshFromGit() throws Exception {
-
-        HashMap<String, Long> timeClaims = new HashMap<>();
-        String token = TokenUtils.generateTokenString("/JwtClaimsWriter.json", timeClaims);
-
-        Engagement engagement = mockEngagement();
-        engagement.setDescription(SCENARIO.SUCCESS.getValue());
-
-        String body = quarkusJsonb.toJson(engagement);
-
-        // POST engagement
-        given()
-            .when()
-                .auth()
-                .oauth2(token)
-                .body(body)
-                .contentType(ContentType.JSON)
-                .post("/engagements")
-            .then()
-                .statusCode(201)
-                .body("customer_name", equalTo(engagement.getCustomerName()))
-                .body("project_name", equalTo(engagement.getProjectName()))
-                .body("project_id", nullValue());
-
-        // GET
-        given()
-            .when()
-                .auth()
-                .oauth2(token)
-                .get("/engagements/customers/" + engagement.getCustomerName() + "/projects/" + engagement.getProjectName())
-            .then()
-                .statusCode(200)
-                .body("customer_name", equalTo(engagement.getCustomerName()))
-                .body("project_name", equalTo(engagement.getProjectName()))
-                .body("project_id", nullValue());
-
-        // Run sync
-        given()
-            .when()
-            .auth()
-            .oauth2(token)
-            .put("/engagements/refresh")
-        .then()
-            .statusCode(200);
-
-        // make sure the async processes finish
-        TimeUnit.SECONDS.sleep(1);
-
-        // GET all engagement
-        Response response = 
-        given()
-            .when()
-                .auth()
-                .oauth2(token)
-                .contentType(ContentType.JSON)
-                .get("/engagements");
-
-        assertEquals(200, response.getStatusCode());
-        Engagement[] engagements = quarkusJsonb.fromJson(response.getBody().asString(), Engagement[].class);
-        assertEquals(1, engagements.length);
-
-        assertEquals(4321, engagements[0].getProjectId());
-        assertEquals("anotherCustomer", engagements[0].getCustomerName());
-        assertEquals("anotherProject", engagements[0].getProjectName());
-
-    }
-
     private Engagement mockEngagement() {
 
         Engagement engagement = Engagement.builder().customerName("TestCustomer").projectName("TestProject")
